qa/deepsea_deploy: implement Stage 1 and Stage 2
Implementation of Stage 1 is complete. Stage 2 is still missing
a lot - notably, non-default storage profiles. Also, currently
the "default profile" deploys OSDs on all eligible disks on
every storage node. This is non-intuitive, because the roles
stanza (in the teutholoy job yaml) expects to specify the
number of osd roles (OSDs) on the individual test nodes.

Signed-off-by: Nathan Cutler <ncutler@suse.com>